Fix caller ghosting after entering call and exchange info
This is the first of three parts to fix the ghosting issue.
Here, one of the root causes of caller ghosting is when the
DxOperator.Patience value was not increased after receiving
the last transmission from user. Thus, it would timeout sooner
than expected. To fix this problem, the Patience value increased
by calling TDxOperator.MorePatience. This is basically saying
that when the DxOperator receives a transmission from the User,
they will continue to work the QSO without loosing patience
and giving up - thus the notion of "more patience" in introduced.

This fixes issue #200. See Issue #200 for more information.