Fix ghosting by stations leaving the QSO early after calling them
This is the second of three parts to fix the ghosting issue.
This fix addresses the second root cause of caller ghosting
where a Dx station would disappear almost immediately after
being created. This would occur about 10% of the time. The user
would enter and send the Dx station's callsign, only to discover
that the station goes silent and does not return its exchange information.

The root cause for this second problem was traced back to a call to the
random function RndRayleigh(4). The result of this call is used to set
the value of DxOperator.Patience. Patience represents operator patience,
an expression of how patient the Dx operator will be while waiting for
a responce from the user. RndRayleigh(4) will return a random integer
value with a Mean of 4. The problem is, 11% of the time, this value
could be 0 or 1. Given a value of 0 or 1, other logic will cause the
DxStation would disappear almost immediately. This was fixed by replacing
the RndRayleight(4) call with '2 + RndRayleigh(3)'. This change makes sure
that a minimum value of 2 is returned. The new Mean is now 5 instead of 4.

This fixes issue #200. See Issue #200 for more information.