Fix problem where ghosted calls are not added to log
This is the third of three changes addressing the issue of
caller ghosting. A ghosted call was originally handled as
a failed QSO after the Dx Station timeout out due to User
inactivity. Once failed, it was deleted from the simulation
before the QSO was logged by the user. With this fix, DxStation
will stop sending and remaining in the set of active stations.
This allows the QSO to be logged as usual by the user.

This fixes issue #200. See Issue #200 for more information.